✨ feat(compat): Add importlib_metadata compatibility
- Introduces a compatibility layer for `importlib.metadata.packages_distributions`.
- Backfills the `packages_distributions` function for Python versions where it's not natively available.
- This ensures consistent behavior across different Python environments by providing a fallback implementation.
